refactored drawingboard to layout
