--- spacebar/src/components/DrawingBoard/index.js
@@ -1,2 +1,0 @@
-import DrawingBoard from "./DrawingBoard";
-export default DrawingBoard;
