+++ spacebar/src/components/Hangout/index.js
@@ -1,0 +1,2 @@
+import Hangout from "./Hangout.jsx";
+export default Hangout;
