--- spacebar/src/pages/DrawingBoardPage.jsx
@@ -1,6 +1,0 @@
-import React from "react";
-import DrawingBoard from "../components/DrawingBoard";
-
-export default function DrawingBoardPage() {
-  return <DrawingBoard />;
-}
