+++ spacebar/src/pages/HangoutPage.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import Hangout from "../components/Hangout";
+
+export default function HangoutPage() {
+  return <Hangout />;
+}
